pyproject.toml
[project]
name = "voiceagent"
version = "0.2.0"
description = "Voice Agent Platform - Build and deploy AI voice agents with data collection"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents[voice]>=0.1.0",
    "groq>=0.11.0",
    "litellm>=1.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
    "rich>=13.0.0",
    "gtts>=2.5.0",
    "scipy>=1.10.0",
    "httpx>=0.27.0",
    "torch>=2.0.0",
    "snac>=1.2.1",
    "einops>=0.8.0",
    "networkx>=3.0",
    "sympy>=1.11",
    # Platform dependencies
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "streamlit>=1.30.0",
    "streamlit-audiorecorder>=0.0.2",
    "nest-asyncio>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

